Move generate to /cli/generator
