+++ cli/generator/handler_test.go
