+++ cli/generator/model.go
