+++ cli/generator/model_test.go
