+++ cli/generator/parse_test.go
