+++ cli/generator/writer.go
